Project 5 - FAQs dropdown
Played around with this tutorial but I could feel that it was getting a bit out of my reach. I knew this because I was aimlessly copying code without actually trying to think and understand what the code actually meant.
Getting a bit hard to see the end goal at this point but going to keep on trying and hopefully something will stick. 
Going to head back to the Bootcamp for a while to begin with ASYNC JS